Break store out of DefaultRelationshipGroupCursor
